// src/app/pages/coordenadas/components/coordenadas-table/coordenadas.interface.ts
export interface CoordenadasInterface {
    claveauth?: string;
    nicknameauth?: string;
    usuarioauth?: string;
    idcoordenada?: number;
    descripcion: string;
    direccion: string;
    medidasterreno: string;
    medidasconstruccion: string;
    fechainicio: string;
    fechafin: string;
    idtipocoordenada: number;
    presupuesto: number;
    idrazonsocialcliente: number;
    idrazonsocialcontratista: number;
    idrazonsocialconstructor: number;
    idrazonsocialasociado: number;
    posiciongps: string;
    idestatuscoordenada: number;
    observaciones: string;
    desctipocoordenada?: string;
    direccioncliente?: string;
    razonsocialcontratista?: string;
    razonsocialconstructor?: string;
    razonsocialasociado?: string;
    claveestatuscoordenada?: string;
    tipocoordenada?: string;
}